+++ tests/Process/Process/GetParentPrivatePropertyValueTest.php
@@ -7,15 +7,13 @@ namespace Steevanb\ParallelProcess\Tests\Process\Process;
 use PHPUnit\Framework\TestCase;
 use Steevanb\ParallelProcess\{
     Process\Process,
-    Tests\CreateLsProcessTrait,
-    Tests\GetReflectionClosureTrait
+    Tests\CreateLsProcessTrait
 };
 
 /** @covers \Steevanb\ParallelProcess\Process\Process::getParentPrivatePropertyValue */
 final class GetParentPrivatePropertyValueTest extends TestCase
 {
     use CreateLsProcessTrait;
-    use GetReflectionClosureTrait;
 
     public function testVisibility(): void
     {
@@ -28,7 +26,7 @@ final class GetParentPrivatePropertyValueTest extends TestCase
         $reflection = $this->createReflection($process);
         $reflection->setAccessible(true);
 
-        static::assertSame('ready', $this->getReflectionClosure($reflection, $process)->call($process, 'status'));
+        static::assertSame('ready', $reflection->getClosure($process)->call($process, 'status'));
     }
 
     public function testStarttime(): void
@@ -40,7 +38,7 @@ final class GetParentPrivatePropertyValueTest extends TestCase
 
         static::assertLessThanOrEqual(
             microtime(true),
-            $this->getReflectionClosure($reflection, $process)->call($process, 'starttime')
+            $reflection->getClosure($process)->call($process, 'starttime')
         );
     }